[refactor] extract ml_inference and qa modules from server.py
